requirements.txt
asgiref==3.2.10
autopep8==1.5.4
beautifulsoup4==4.9.3
confusable-homoglyphs==3.2.0
Django==3.1.2
django-bootstrap4==2.3.1
django-crispy-forms==1.10.0
django-registration==3.1.1
django-rest-authtoken==2.1.3
django-tinymce==3.2.0
djangorestframework==3.12.2
numpy==1.19.5
Pillow==8.1.0
psycopg2-binary==2.8.6
pycodestyle==2.6.0
python-decouple==3.4
pytz==2020.5
soupsieve==2.1
sqlparse==0.4.1
toml==0.10.2
